backend/requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
waitress
emoji
flask-cors
gunicorn
Flask-Mail
Flask-APScheduler
python-dotenv
celery
redis
pyjwt[crypto]
requests